requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via starlette
beautifulsoup4==4.13.4
    # via -r requirements.in
browserforge==1.2.3
    # via -r requirements.in
certifi==2025.1.31
    # via requests
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via
    #   browserforge
    #   uvicorn
colorama==0.4.6
    # via click
fastapi==0.115.12
    # via -r requirements.in
h11==0.16.0
    # via uvicorn
idna==3.10
    # via
    #   anyio
    #   requests
pydantic==2.11.3
    # via fastapi
pydantic-core==2.33.1
    # via pydantic
python-decouple==3.8
    # via -r requirements.in
requests==2.32.3
    # via -r requirements.in
sniffio==1.3.1
    # via anyio
soupsieve==2.7
    # via beautifulsoup4
starlette==0.46.2
    # via fastapi
typing-extensions==4.13.2
    # via
    #   beautifulsoup4
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.0
    # via pydantic
urllib3==2.4.0
    # via requests
uvicorn==0.34.2
    # via -r requirements.in
